# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "DeleteResponse": {
        "additionalProperties": false,
        "properties": {
          "deletedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deletedat"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "DeleteResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SharedTimetableRead": {
        "additionalProperties": false,
        "properties": {
          "metrics": {
            "$ref": "#/components/schemas/TimetableMetrics"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "title": "Sections",
            "type": "array"
          },
          "timetable": {
            "$ref": "#/components/schemas/TimetableRead"
          }
        },
        "required": [
          "timetable",
          "sections",
          "metrics"
        ],
        "title": "SharedTimetableRead",
        "type": "object"
      },
      "TimetableCopyRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "TimetableCopyRequest",
        "type": "object"
      },
...
      "TimetableDetail": {
        "additionalProperties": false,
        "properties": {
          "conflictSectionIds": {
            "items": {
              "maxItems": 2,
              "minItems": 2,
              "prefixItems": [
                {
                  "type": "string"
                },
                {
                  "type": "string"
                }
              ],
              "type": "array"
            },
            "title": "Conflictsectionids",
            "type": "array"
          },
          "metrics": {
            "$ref": "#/components/schemas/TimetableMetrics"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/Section"
            },
            "title": "Sections",
            "type": "array"
          },
          "timetable": {
            "$ref": "#/components/schemas/TimetableRead"
          }
        },
        "required": [
          "timetable",
          "sections",
          "metrics",
          "conflictSectionIds"
        ],
        "title": "TimetableDetail",
        "type": "object"
      },
...
      "TimetableList": {
        "additionalProperties": false,
        "properties": {
          "timetables": {
            "items": {
              "$ref": "#/components/schemas/TimetableRead"
            },
            "title": "Timetables",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "timetables",
          "total"
        ],
        "title": "TimetableList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "PL-timeTabler API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/shared-timetables/{share_code}": {
      "get": {
        "operationId": "get_shared_timetable_api_v1_shared_timetables__share_code__get",
        "parameters": [
          {
            "in": "path",
            "name": "share_code",
            "required": true,
            "schema": {
              "title": "Share Code",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SharedTimetableRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Shared Timetable",
        "tags": [
          "shared-timetables"
        ]
      }
    },
    "/api/v1/shared-timetables/{share_code}/copy": {
      "post": {
        "operationId": "copy_shared_timetable_api_v1_shared_timetables__share_code__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "share_code",
            "required": true,
            "schema": {
              "title": "Share Code",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableCopyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy Shared Timetable",
        "tags": [
          "shared-timetables"
        ]
      }
    },
    "/api/v1/timetables": {
      "get": {
        "operationId": "list_timetables_api_v1_timetables_get",
        "parameters": [
          {
            "in": "query",
            "name": "semester",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester"
            }
          },
          {
            "in": "query",
            "name": "favorite",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Favorite"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Timetables",
        "tags": [
          "timetables"
        ]
      },
      "post": {
        "operationId": "create_timetable_api_v1_timetables_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Timetable",
        "tags": [
          "timetables"
        ]
      }
    },
    "/api/v1/timetables/history": {
      "get": {
        "operationId": "timetable_history_api_v1_timetables_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "semester",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Semester"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Timetable History",
        "tags": [
          "timetables"
        ]
      }
    },
    "/api/v1/timetables/{timetable_id}": {
      "delete": {
        "operationId": "delete_timetable_api_v1_timetables__timetable_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Timetable",
        "tags": [
          "timetables"
        ]
      },
      "get": {
        "operationId": "get_timetable_api_v1_timetables__timetable_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Timetable",
        "tags": [
          "timetables"
        ]
      },
      "patch": {
        "operationId": "update_timetable_api_v1_timetables__timetable_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Timetable",
        "tags": [
          "timetables"
        ]
      }
    },
    "/api/v1/timetables/{timetable_id}/copy": {
      "post": {
        "operationId": "copy_timetable_api_v1_timetables__timetable_id__copy_post",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableCopyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Copy Timetable",
        "tags": [
          "timetables"
        ]
      }
    },
    "/api/v1/timetables/{timetable_id}/favorite": {
      "patch": {
        "operationId": "update_favorite_api_v1_timetables__timetable_id__favorite_patch",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableFavoriteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Favorite",
        "tags": [
          "timetables"
        ]
      }
    },
    "/api/v1/timetables/{timetable_id}/sections": {
      "patch": {
        "operationId": "update_timetable_sections_api_v1_timetables__timetable_id__sections_patch",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableItemsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Timetable Sections",
        "tags": [
          "timetables"
        ]
      }
    },
    "/api/v1/timetables/{timetable_id}/shares": {
      "post": {
        "operationId": "create_share_api_v1_timetables__timetable_id__shares_post",
        "parameters": [
          {
            "in": "path",
            "name": "timetable_id",
            "required": true,
            "schema": {
              "title": "Timetable Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TimetableShareCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TimetableShareRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Share",
        "tags": [
          "timetables"
        ]
      }
    },
...
  }
}